Changed WAL record dispatch method from shm+socket to shm+mqueue.
Here, we use POSIX message queue.   It is simpler, atomic and fast.

We need to consider that POSIX mqueue configuration is just done
in OS level.   For example, mqueue is located on /dev/mqueue.
In addition, number of total mqueue, number of queue, total amount
of memory for total queued data are all subject to OS configurartion
and default setting looks minimum.

Good thing is all the flow control is embedded in mqueue, API is verr
simple, faster than socket.

	modified:   src/backend/access/transam/parallel_replay.c
	modified:   src/backend/access/transam/xlog.c
	modified:   src/include/access/parallel_replay.h